LPS-109717 Move useDragAndDrop to utils
